login/loginCheck.go
package login

import "database/sql"
import "PillarsFlowNet/utility"

func QueryUserCode(userName * string, password * string, db * sql.DB) * string {
	stmt, err := db.Prepare("SELECT user_code FROM user WHERE user_name=? AND password=?")
	if err != nil {
		panic(err.Error())
	}
	defer stmt.Close()
	passwordMd5 := utility.Md5sum(password)
	result, err := stmt.Query(userName, passwordMd5)
	if err != nil {
		panic(err.Error())
	}
	defer result.Close()
	var user_code string

	if result.Next() {
		result.Scan(&user_code)
		
	} 
	return &user_code
}

